Fix MPI errors (follow-up to bace230)

+++ parallel/src/sgpp/parallel/finance/application/BlackScholesSolverMPI.cpp
@@ -122,9 +122,9 @@ void BlackScholesSolverMPI::getGridNormalDistribution(sgpp::base::DataVector& al
   }
 }
 
-void BlackScholesSolverMPI::constructGrid(sgpp::base::BoundingBox& BoundingBox, int level) {
+void BlackScholesSolverMPI::constructGrid(sgpp::base::BoundingBox& BoundingBox, size_t level) {
   this->dim = BoundingBox.getDimensions();
-  this->levels = level;
+  this->levels = static_cast<int>(level);
 
   this->myGrid = new base::LinearBoundaryGrid(BoundingBox);
